Drop all support for creating the Delivery Client (#11)
Going to be a bit opinionated about Delivery client creation here. Using DI to inject the client is better than doing it inline in the pipeline.
Additionally, this removes a bunch of code that is hard to test and sort-of duplicates what the Delivery SDK already offers.
The downside is that the initial setup is a bit harder.